Move app.jsx out to get HMR
